+++ app/features/pengajar/constants.ts
@@ -1,5 +1,7 @@
 import { UAvatar } from "#components";
 import type { TableColumn } from "@nuxt/ui";
+import { z } from "zod/mini";
+import { genderEnum, statusPengajarEnum } from "~~/shared/enum";
 
 export const columns: TableColumn<any>[] = [
   {
@@ -33,3 +35,40 @@ export const columns: TableColumn<any>[] = [
     header: "No. Telepon",
   },
 ];
+
+export const schema = z.object({
+  id: z.optional(z.number()),
+  foto: z.string(),
+  file: z.optional(
+    z
+      .file()
+      .check(
+        z.maxSize(5_000_000),
+        z.mime(["image/png", "image/jpeg", "image/webp"])
+      )
+  ),
+  gender: z.enum(genderEnum),
+  nama: z.string().check(z.minLength(1, "Required")),
+  noTelepon: z.string().check(z.minLength(1, "Required")),
+  pendidikan: z.string().check(z.minLength(1, "Required")),
+  status: z.enum(statusPengajarEnum),
+  tanggalLahir: z.nullable(z.string().check(z.minLength(1, "Required"))),
+  tempatLahir: z.string().check(z.minLength(1, "Required")),
+  tanggalTugas: z.nullable(z.string().check(z.minLength(1, "Required"))),
+});
+
+export const getInitialFormData = (): Schema => ({
+  id: undefined,
+  nama: "",
+  gender: "Laki-laki",
+  noTelepon: "",
+  pendidikan: "",
+  status: "Asisten Pengajar",
+  tanggalLahir: "",
+  tempatLahir: "",
+  tanggalTugas: "",
+  foto: "",
+  file: undefined,
+});
+
+export type Schema = z.infer<typeof schema>;
