+++ app/features/pengajian/absensi/components/EmptyCard.vue
@@ -1,0 +1,11 @@
+<template>
+  <UCard>
+    <div class="flex flex-col items-center justify-center gap-4 py-16">
+      <UIcon name="i-lucide-user-check" size="50" />
+      <h2 class="text-xl font-bold">Pilih Pengajian</h2>
+      <p class="text-muted text-center">
+        Silahkan pilih pengajian di menu diatas untuk memulai absensi
+      </p>
+    </div>
+  </UCard>
+</template>
