+++ app/features/pengajian/absensi/types.ts
@@ -1,0 +1,11 @@
+export type QueryType = Partial<{
+  page: number;
+  search: string;
+}>;
+
+export type DataReturn = {
+  id: number;
+  jamaahId: number;
+  detail: string;
+  keterangan: "Hadir" | "Izin" | "Tanpa Keterangan";
+};
