+++ app/features/pengajian/jamaah/types.ts
@@ -1,0 +1,9 @@
+export type QueryType = Partial<{
+  page: number;
+  search: string;
+}>;
+
+export type DataReturn = {
+  id: number;
+  nama: string;
+};
