+++ app/features/pengajian/monitoring/JamaahMonitoringContainer.vue
@@ -1,22 +1,14 @@
 <script setup lang="ts">
-  import { useConstantStore } from "~/stores/constant";
-  import { columns } from "./_constants";
-  import { APIBASE } from "~/utils";
-
-  const constantStore = useConstantStore();
-  constantStore.setTitle("PJP Kelompok / Monitoring Kehadiran Pengajian");
+  import { columns } from "./constants";
+  import type { QueryType } from "./types";
 
   const { data: summary } = await useFetch(
     `${APIBASE}/absensi-jamaah/monitoring/summary`
   );
 
-  const query = reactive({
-    search: "",
+  const query = reactive<QueryType>({
     page: 1,
   });
-  const searchDebounced = useDebounceFn((v) => {
-    query.search = v;
-  }, 300);
   const { data, status } = await useFetch(
     `${APIBASE}/absensi-jamaah/monitoring`,
     {
@@ -26,18 +18,17 @@
 </script>
 
 <template>
-  <Title>PJP Kelompok | Monitoring Kehadiran Pengajian</Title>
   <main class="flex flex-col gap-4">
-    <div class="grid grid-cols-1 gap-4 md:grid-cols-2">
+    <div class="grid grid-cols-2 gap-4">
       <UCard>
-        <p class="flex items-center gap-4 text-3xl font-bold md:text-4xl">
+        <p class="flex items-center gap-4 text-4xl font-bold">
           <UIcon name="i-lucide-users" />
           {{ summary?.data.countJamaah }}
         </p>
         <p class="text-muted">Total Jamaah</p>
       </UCard>
       <UCard>
-        <p class="flex items-center gap-4 text-3xl font-bold md:text-4xl">
+        <p class="flex items-center gap-4 text-4xl font-bold">
           <UIcon name="i-lucide-trending-up" />
           {{ summary?.data.kehadiran }}%
         </p>
@@ -46,13 +37,7 @@
     </div>
     <UCard>
       <div class="mb-4 flex gap-2 md:mb-6 md:gap-4">
-        <UInput
-          size="xl"
-          class="flex-5"
-          leading-icon="i-lucide-search"
-          placeholder="Search..."
-          @update:model-value="searchDebounced"
-        />
+        <InputSearch v-model="query.search" />
       </div>
       <AppTable
         v-model:page="query.page"
