+++ app/features/pengajian/monitoring/types.ts
@@ -1,0 +1,4 @@
+export type QueryType = Partial<{
+  page: number;
+  search: string;
+}>;
