+++ app/pages/dashboard/(PJP Kelompok)/kelompok-absensi-pengajian.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+  import AbsensiJamaahContainer from "~/features/pengajian/absensi/AbsensiJamaahContainer.vue";
+  import { useConstantStore } from "~/stores/constant";
+
+  const constantStore = useConstantStore();
+  constantStore.setTitle("PJP Kelompok / Absensi Jamaah");
+</script>
+
+<template>
+  <Title>PJP Kelompok | Absensi Jamaah</Title>
+  <AbsensiJamaahContainer />
+</template>
